Move storage_contents/1 and files_on_storage/1 to StorageQuery
